+++ packages/components/modules/profiles/common/graphql/mutations/ProfileUserRoleCreate.ts
@@ -6,12 +6,17 @@ import { Disposable, UseMutationConfig, graphql, useMutation } from 'react-relay
 import { ProfileUserRoleCreateMutation } from '../../../../../__generated__/ProfileUserRoleCreateMutation.graphql'
 
 export const ProfileUserRoleCreateMutationQuery = graphql`
-  mutation ProfileUserRoleCreateMutation($input: ProfileUserRoleCreateInput!) {
+  mutation ProfileUserRoleCreateMutation(
+    $input: ProfileUserRoleCreateInput!
+    $connections: [ID!]!
+  ) {
     profileUserRoleCreate(input: $input) {
-      profileUserRoles {
+      profileUserRoles
+        @prependNode(connections: $connections, edgeTypeName: "ProfileUserRoleEdge") {
         id
         status
         role
+        ...MemberItemFragment
       }
       errors {
         field